+++ module/civoShell.psm1
@@ -1,6 +1,13 @@
-Get-ChildItem -Name "*.ps1" -Path ($PSScriptRoot + "/functions") | ForEach-Object {
-    . ($PSScriptRoot + "/functions/" + $_)
-}
-Get-ChildItem -Name "*.ps1" -Path ($PSScriptRoot + "/functions/kubernetes") | ForEach-Object {
-    . ($PSScriptRoot + "/functions/kubernetes/" + $_)
+# ------------- civoShell ------------- #
+# Dot sourcing all available functions. #
+$functionTypes = @(
+    '/'
+    '/kubernetes/'
+    '/networking/'
+)
+
+foreach ($function in $functionTypes) {
+    Get-ChildItem -Name "*.ps1" -Path ($PSScriptRoot + "/functions" + $function) | ForEach-Object {
+        . ($PSScriptRoot + "/functions" + $function + $_)
+    }
 }