+++ module/functions/kubernetes/New-CivoKubernetesCluster.ps1
@@ -1,4 +1,40 @@
 function New-CivoKubernetesCluster {
+    <#
+    .SYNOPSIS
+    Create a new Kubernetes Cluster.
+    .DESCRIPTION
+    Create a new k3s Kubernetes cluster.
+    You can either just define the name of the cluster or define it just as you want. See the available parameters for what you can do.
+    .PARAMETER Name
+    Set the name of your cluster. This parameter is mandatory and must be set to continue.
+    .PARAMETER NodeCount
+    Set the number of nodes you want to run. The default is 3.
+    .PARAMETER NodesSize
+    Set the instance size of your nodes. The current default is g2.small.
+    .PARAMETER KubernetesVersion
+    Set the version of Kubernetes you want to run. This defaults to the latest available.
+    .PARAMETER Tags
+    You can define one or more tags for your cluster. The list should be space separated.
+    .PARAMETER Applications
+    A comma separated list of applications you want install at deployment. If you want to remove a default application, you can prefix the name with a -.
+    .OUTPUTS
+    Details of the new cluster you have created.
+    .EXAMPLE
+    New-CivoKubernetesCluster MyNewCluster
+    .EXAMPLE
+    New-CivoKubernetesCluster -Name "Cluster01" -NodeCount 1
+    .EXAMPLE
+    $cluster            = @{
+        Name            = "ClustersLastStand"
+        NodeCount       = 2
+        NodesSize        = "g2.large"
+        Tags            = "prod powershell"
+        Applications    = "prometheus-operator, MinIO"
+    }
+    New-CivoKubernetesCluster @cluster
+    .LINK
+    https://github.com/roberthstrand/civoShell
+    #>
     [CmdletBinding()]
     param (
         [Parameter(Position = 0, Mandatory)]
@@ -6,10 +42,10 @@ function New-CivoKubernetesCluster {
         $Name,
         [Parameter(Position = 1)]
         [string]
-        $NodeCount = 3,
+        $NodeCount,
         [Parameter(Position = 2)]
         [string]
-        $NodeSize,
+        $NodesSize,
         [Parameter(Position = 3)]
         [string]
         $KubernetesVersion,
@@ -23,7 +59,7 @@ function New-CivoKubernetesCluster {
     $Form                   = @{
         name                = $Name
         num_target_nodes    = $NodeCount
-        target_node_size    = $NodeSize
+        target_nodes_size   = $NodesSize
         kubernetes_version  = $KubernetesVersion
         tags                = $Tags
         applications        = $Applications
@@ -41,7 +77,9 @@ function New-CivoKubernetesCluster {
         Status              = $call.status
         Ready               = $call.ready
         Nodes               = $call.num_target_nodes
+        NodesSize           = $call.target_nodes_size
         KubernetesVersion   = $call.kubernetes_version
         Tags                = $call.tags
+        Applications        = $call.installed_applications.application
     }
 }